megarepo_api: Derive data remotely - behind a jk
Summary:
Replace the calls to `derive_bulk_locally` with `derive_bulk`.

This should address the issues we are seeing of large requests to `megarepo_api` getting stuck in OOM crashloops.

It's a slightly risky change as derivation may slow down in the common case, causing AOSP facing SEVs.
For that reason, we gate the change behind a jk and will rollout through that jk while monitoring prod usage.

Reviewed By: andreacampi

Differential Revision: D77799476

fbshipit-source-id: 779ce5c149b566aa31aa1e54a0378066bfce5267