Rename `derive_bulk` to `derive_bulk_locally`
Summary:
One of the specificities of this method is that it enforces local derivation.

If we instead call `derive`, the `local`/`remote` question is more nuanced:
It will derive remotely if so configured and fallback to local derivation if necessary.

I want to introduce a new variant of `derive_bulk` that has the remote semantics of `derive`.
Instead of naming that new method: `derive_bulk_remotely_if_so_configured_and_fallback_to_local_derivation_if_necessary`,
I'm going for renaming this one to `derive_bulk_locally`, freeing the name for my new favourite child.

Reviewed By: YousefSalama

Differential Revision: D77799474

fbshipit-source-id: 652809e36de43b038269c505cef318bed7e37a3b