megarepo_api: Use special bucket
Summary:
If the repo is configured to have a derived data config named "megarepo_api_rollout_remote_derivation", use it.

In conjunction with D77798137, this will allow us to isolate derivation requests caused by this code change,
enabling us to mitigate any consequences of an overly eager megarepo_api in prod faster than if we had filled
the default derivation queue with millions of requests.

Reviewed By: andreacampi

Differential Revision: D77799473

fbshipit-source-id: d1590a5cbc0de062e0af4bbe18b09b555dbc930c