Rename SqlTelemetryLogger to SqlQueryTelemetry
Summary:
## This stack
Previous diffs focused on piping the stats returned by the MySQL client up to mononoke_queries.

The next few diffs will focus on changing mononoke_queries and the mononoke codebase to start logging the data.

## This diff
`SqlTelemetryLogger` as not a great name... I named it this orignally because the initial implementation would store a scuba logger, but that's not the case anymore...
Since its main function is to store data to be logged, let's rename it.

Reviewed By: andreacampi

Differential Revision: D77787233

fbshipit-source-id: b669a9cbf7527a5c219affe57a1874be1b3601db