crdump: fix faulty reporting of "commit_cloud"
Summary:
Previously if a commit failed to upload to commit cloud in a certain way, we would still report the commit as "commit_cloud: true" in the response. This was because the commit cloud upload() method returns a set of failed commit nodes, but the "commit_cloud" check tries indexing using commit revision. Fix by converting the nodes back to revisions (nodes are better, but this was more convenient).

The upload failure seems to be when _something_ happens on the server during changeset upload, and the server  doesn't return some or all of the commits. The upload code notices this and reports the commits that didn't upload, but unfortunately the actual error is not propagated, so we need some more investigation to understand why the commit failed to upload.

Reviewed By: zzl0

Differential Revision: D77766836

fbshipit-source-id: 7114946fca2314a4b50361800c2ae73675a73b6a