Log mononoke request metadata to mononoke_sql_telemetry
Summary:
## This stack
Previous diffs focused on piping the stats returned by the MySQL client up to mononoke_queries.

The next few diffs will focus on changing mononoke_queries and the mononoke codebase to start logging the data.

## This diff
As title. This will log data that's useful for debugging.

markbt I see that the Metadata crate is under `eden/mononoke/server`, so I wasn't sure if we could introduce the dependency because it could break SCS.

However, I checked that it has only a few deps itself **and that the [`scs_methods` crate](https://fburl.com/code/qc6ho7fc) already depends on it, so it should be fine

Reviewed By: markbt

Differential Revision: D77861736

fbshipit-source-id: 5774b2198ac9b8a52d01a2196d231dd8c7b87322