stream changes since with states - add states check
Summary:
# Context

We are introducing EdenFS notifications to support scalable and ergonomic file system notifications for EdenFS mounts.

# This Stack
Implement state enter/state leave. See design doc.

# This diff
Adds framework for handling reading states.

Adds new function check_states_asserted. This function returns a bool. False means states are not asserted. True means that at least 1 state is asserted.

Add demo behavior based on states.

If the result is false, continue to reading the inner stream.
if the result is true, it waits 1 second then checks again.
If the result is an error, return None, ending the stream.

# Technical Details
KNOWN ISSUES:
Right now there's a leak on the initial check, since next() is being waited on indefinitely, the first event will go through regardless of the states being asserted. This will be addressed in a followup where we inject the states as events into the stream.

# Discussion Points

Reviewed By: muirdm

Differential Revision: D73813444

fbshipit-source-id: efa180ef3a684efd8783705b541cec2294e86d38