Remove blobimport from targets that do not use it
Summary:
A number of integration test list blobimport as a dependency when they in fact do not use it.

This is not caught at land time because of the way our test runner is set, where we set up environment variables to point to targets, but wthe environment varibales might not be used.

I am also setting up a butterfly rule to prevent more blobimport as dependencies being added: https://www.internalfb.com/butterfly/rule/1281619930272603

Reviewed By: RajivTS

Differential Revision: D77867484

fbshipit-source-id: 3895956043759ef3eeabf79afdeb7d9290384494