Log query errors to mononoke_sql_telemetry
Summary:
## This stack
Previous diffs focused on piping the stats returned by the MySQL client up to mononoke_queries.

The next few diffs will focus on changing mononoke_queries and the mononoke codebase to start logging the data.

## This diff

It will be useful to log any exceptions from the queries themselves to the table as well. We'll likely be running queries to analyse failures and it's useful to have the errors visible in the SC Correlator Dashboard (which I already updated to have the new table BTW)

Reviewed By: YousefSalama

Differential Revision: D77881094

fbshipit-source-id: 243677519e04f726b1da29c9d97e9055f8d469f8