Use existing function to log client_request_info
Summary:
`client_request_info` is in `metadata`. There's a function to add it to the Scuba sample, yet `add_metadata` has the same exact same code.

Let's re-use the helper to remove the duplication **and ensure that the IDs of blobstores with disabled reads are always logged** (not happening in `add_metadata`)

Reviewed By: YousefSalama

Differential Revision: D77864122

fbshipit-source-id: b29cb4deb909035689e36f7de21ad1a5bdb692c3