Add a ClientEntryPoint for tests
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on testing.

## This diff
As title. I prefer having one for tests over using `Unknown`.
I went ahead and started using in the test binary I added in D77866823 as well.

Reviewed By: zzl0

Differential Revision: D77939514

fbshipit-source-id: 7d16435479b85ac2458d53acc53ab459b6267ff9