Remove lock acquisition on tailer startup
Summary:
Initially I added a lock to ensure that we were only running on one shard at a time per region. This was based on some wrong assumptions:
- That shard manager would not ensure that at least one replica is running. In reality, given enough capacity, shard manager will move the replicas if necessary to ensure they are up.
- That derived data tailer must run ALWAYS. In reality we can accept a few (single digit) minutes without it running. I've double checked and a task restart for derived data tailer is roughtly 1 minute.
Finally, if restart time turns out to be unacceptable, we think the additional load on the blobstores from running 2 replicas per repo will not be excessive.

Reviewed By: andreacampi

Differential Revision: D77938231

fbshipit-source-id: 83cee3eadf1681886b8837fc9539f62c88e5abba