Add mocked constructor for SqlQueryConfig
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on testing.

## This diff
I started adding it to the test itself, but IMO it makes more sense for it to be here, so it can be used by any other tests in the future.

Reviewed By: andreacampi

Differential Revision: D77939513

fbshipit-source-id: b2b64c450dd5fdaf2661491d19e3d3eb5970f375