Move imports to the test module scope
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on testing.

## This diff
Instead of having them in the single unit test that doesn't actually run (just checks if it compiles).

These imports will be used by the new unit tests in future diffs.

Reviewed By: andreacampi

Differential Revision: D77939516

fbshipit-source-id: 2bbd42300544958177cc7a3b058ccddc1555c983