Add mode parameter to open-isl-with-commit-message command
Summary:
Extends the `sapling.open-isl-with-commit-message` VS Code command to support both commit and amend modes. The command now accepts an optional third parameter `mode` that can be 'commit' or 'amend', allowing callers to specify whether the ISL should be opened in commit mode (default) or amend mode.

Also allowing consumers to update or add a new commit for some changes.

Reviewed By: stbfb

Differential Revision: D77960811

fbshipit-source-id: 7335b30a9bb0ff6cbd9fc021bd0bc71e69a95f0e