Early exit on error from subscription
Summary:
This Diff:

When stream_changes_since hits an error, terminate the stream.
This is done at two locations, once if the inner subscription returns an error, and once if there is an error when calling get_changes_since.
In the second case, return that error, then terminate the stream.

Reviewed By: muirdm

Differential Revision: D77717934

fbshipit-source-id: 1af8167f60e1c721d1fb24ef8f586e4271f5509f