Get repo_id from metadata XDB queries
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on testing.

## This diff
One of the most important things we want in the telemetry is to log repos involved in a query **or transaction**. To do that, we need to log the ids or names of the repos.

There are multiple ways we could do that, one of which would involve changing the mononoke_queries API to allow any queries that involve repos, to provide that.

Another way would be to extract that automatically from the query parameter names and/or types.

None of these are exhaustive, i.e. would ensure that all queries that use repo names or ids are having those logged.

I decided to try out option 2 to start getting some logs, but this is definitely not the final solution.

What I'll do next is **start logging the name of the query (e.g. `InsertMapping`, `AddChangesets`) along with its args**.

I'll then go over the data to get a sense of what arguments are used in all queries through scuba, instead of trying to find it all in the code.

After deciding exactly the args I want to keep track of, which can appear in different formats (e.g. repo, repo_name, source_repo_id), I'll make a decision on the final approach to get these values and log them.

Reviewed By: markbt

Differential Revision: D77871410

fbshipit-source-id: 6a89fc878a7a627597efbf239cd7bcf77a5373f1