Add unit tests to mononoke_queries
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on testing.

## This diff
mononoke_queries doesn't have any unit tests right now. The only one is there just to ensure the code compiles, but it's not actually runnable.

This diff creates the first unit test to cover the Scuba logging logic.
It will ensure that the expected values are always logged, with other minor checks, e.g. values that are always supposed to be logged or always the same.

I'll add more unit tests as I make more changes to the `mononoke_queries` macros for logging.

Reviewed By: YousefSalama

Differential Revision: D77935148

fbshipit-source-id: 207295f6573ba0ecd4434e7d22e25405ec2a4bfc