Introduce Transaction wrapper in mononoke's sql_ext crate
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on tracking Transaction level telemetry and getting repo_ids from the queries.

## This diff

Introduces a wrapper over the `sql::Transaction` struct (I'll refer to it as a "Mononoke Transaction"), which we'll use to keep track of the metadata we want to track over a transaction. This will include initially:
- Repos involved in the transaction (e.g. repo_ids, repo_names)
- read and write tables

This diff is a big big, but mostly just replaces all callsites to use the new wrapper. I added comments on the important changes.

My process was:

- Introduce the wrapper with the same public methods (e.g. commit, rollback).
- Add a `From` implementation to cast a `sql::Transaction` into a mononoke transaction.
- Change all the imports to import the Mononoke transaction
- Call `.into()` in all callsites needed to convert between them.
˙mononoke_quer

**How we'll use it**
- Before calling `commented_query` from the rust/shed sql crate, we unwrap the transaction and get the sql txn.
- Call it, get a new txn and wrap it up again.

In the next few diffs I'll add the telemetry to keep track of, so we'll unwrap everything, update the telemetry and wrap it back up.

Reviewed By: YousefSalama

Differential Revision: D77998324

fbshipit-source-id: 5f38790bd7222f11a49bf7302fe4ffbc3db45bf6