Set scuba sample rate
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on tracking Transaction level telemetry and getting repo_ids from the queries.

## This diff

As expected, there is a lot of telemetry being logged, so the Scuba table has a very low retention... It was 0.5 day, but I increased it to 3 and will now introduce sampling, which should give us ~30 days retention.

Reviewed By: YousefSalama

Differential Revision: D78156319

fbshipit-source-id: 5ad486bfad5bc68e6ff37182bb7b1f4be2382d77