confirm that we don't reschedule a periodic task when the interval changed to zero
Summary:
I am going to stop a periodic task in diff D78054136. I want to confirm that when I changed the interval to zero it stopped and canceled the next run. Looking at the code when the interval changed to zero it cancel the timeout:
https://www.internalfb.com/code/fbsource/[24c95884a153]/fbcode/eden/fs/service/PeriodicTask.cpp?lines=72-75
However, if the periodic task is running, reschedule should handle the interval changes:
https://www.internalfb.com/code/fbsource/[24c95884a153]/fbcode/eden/fs/service/PeriodicTask.cpp?lines=67-70

I want to add `cancelTimeout();` to `reschedule()` therefore we can confirm that it cancel the next timer when interval is zero

Differential Revision: D78042244

fbshipit-source-id: dd1db682b114b21173a60f5196906c37136fb2b3