clean up unused code in NFS GC
Summary:
I setup this threshold when we have ESTALE errors from running NFS GC. The goal was to limit running periodic GC to only high number of inodes. Now that we fixed the ESTALE errors in NFS GC we don't need this threshold anymore.

Cleaning up the code, therefore I can implement an aggressive GC in the next diff

Differential Revision: D78100328

fbshipit-source-id: 9b5f8be7ec57bdbd9ac3af3dbea93831c466f850