Unit test for transaction telemetry
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on tracking Transaction level telemetry and getting repo_ids from the queries.

## This diff

Add a unit test before implementing support for transaction telemetry.

When the feature is implemented, I'll just update the unit test

Reviewed By: andreacampi

Differential Revision: D78069691

fbshipit-source-id: f3719fd544a92f89e182cc04862776ac7d2878f3