slapi: fix capabilities on slapi git
Summary:
Slapi-git is a bit of a half-made thing where it technically works for some endpoints (which is why we have tests for it) but in prodution there's not real service that supports it since we don't run mononoke server with git repos and we haven't settled on a formal effort to do so.

This diff basically makes the `capabilities` endpoint on mononoke server show that there's no type of requests that are currently supported for git repos on mononoke server. Note that for mononoke git server we do support commit cloud and the capabilities endpoint there [reflects that behaviour](https://www.internalfb.com/code/fbsource/[87711610b466]/fbcode/scm/metagit/metagit-location-service/src/service.rs?lines=47-54).

Reviewed By: andreacampi

Differential Revision: D77798094

fbshipit-source-id: 360e26d27dac51bf9d8bb0aa279719294f01edb4