Add a new field oldEdenPid_ to EdenServer
Summary:
### This diff
One corner case that we should consider is during graceful restart that we have two instances of eden at the same time. When the new eden is starting up the old eden didn't clean up yet, and there is a possibility that new eden instance find the old eden heartbeat file during start up. To avoid these cases we can keep the PID of old eden, therefore when the new eden starts and finds the old eden heartbeat it will ignore it.

### Context
We found some cases that eden crash silently due to OOM kills more details in S530151.
To understand if any of our fixes are helping, we need a method for tracking when EdenFS is killed unexpectedly (i.e. SIGKILL, OOM, etc). The easiest way to do this is to have EdenFS set a flag file (called eden heartbeat) on startup. On normal exits, EdenFS should clear that flag file. On unexpected exits (crashes, OOMs, SIGKILLs), EdenFS won't clear the flag and we can log that during the next startup. We need to update the latest timestamp in this hearbeat file while eden is alive. Therefore, when the crash gets catch, we can see the last time that eden was alive.

Reviewed By: MichaelCuevas

Differential Revision: D77262540

fbshipit-source-id: 88c05279699122395ea93fa2f5bf194c11bca7f6