Create a config to update eden heartbeat
Summary:
### This diff
Create a config for updating the latest timestamp for eden heartbeat

### Context
We found some cases that eden crash silently due to OOM kills more details in S530151.
To understand if any of our fixes are helping, we need a method for tracking when EdenFS is killed unexpectedly (i.e. SIGKILL, OOM, etc). The easiest way to do this is to have EdenFS set a flag file (called eden heartbeat) on startup. On normal exits, EdenFS should clear that flag file. On unexpected exits (crashes, OOMs, SIGKILLs), EdenFS won't clear the flag and we can log that during the next startup. We need to update the latest timestamp in this hearbeat file while eden is alive. Therefore, when the crash gets catch, we can see the last time that eden was alive.

Reviewed By: MichaelCuevas

Differential Revision: D77963264

fbshipit-source-id: fdcb871550c9a1090707c8289bebc3cba7e9eb13