Remove heartbeat file when SIGKILL signal triggered by eden
Summary:
### This diff
There are some cases that eden triggered the `SIGKILL` signal to force daemon exit. These `SIGKILL`s are not triggered by the OS due to memory issues, so we should clean up the heartbeat file before killing edenFS daemon. This ensures that the `SIGKILL` won't be mislogged as a silent exit when the next EdenFS daemon starts.

### Context
We found some cases that eden crash silently due to OOM kills more details in S530151.
To understand if any of our fixes are helping, we need a method for tracking when EdenFS is killed unexpectedly (i.e. SIGKILL, OOM, etc). The easiest way to do this is to have EdenFS set a flag file (called eden heartbeat) on startup. On normal exits, EdenFS should clear that flag file. On unexpected exits (crashes, OOMs, SIGKILLs), EdenFS won't clear the flag and we can log that during the next startup. We need to update the latest timestamp in this hearbeat file while eden is alive. Therefore, when the crash gets catch, we can see the last time that eden was alive.

Reviewed By: MichaelCuevas

Differential Revision: D78183705

fbshipit-source-id: 62c5fba16e393dfa5d7ae35c2c2c5de443fd9bf0