walkdetector: start tracking recursive directory metadata
Summary:
Previously we only tracked directory size metadata for individual directories. Now we also tabulate the "recursive" totals per-depth for ancestor nodes.

I replaced the previous `WalkNode.{total,file}_dirs` with a vector where the vector index is the depth (and the value is totals from descendants at that depth relative to WalkNode). The previous `WalkNode.{total,file}_dirs` fields are equivalent to index=0 of the new vectors.

I'm going to use this to avoid prematurely advancing a walk boundary to depth+1 when depth+1 is very large.

Reviewed By: zzl0

Differential Revision: D78272867

fbshipit-source-id: 0204e1f63d6394a4b3d46c2329dfcf9b858ad85a