walkdetector: clean up an obsolete check
Summary: This check and test are obsolete now. When merging nodes into parent or grandparent node, we always set depth to 1 or 2, respectively (previously we would set to the max of the nodes being merged, which could be arbitrarily large). It is no longer possible for the parent/grandparent walk to accidentally advance "too far" during the merge.

Reviewed By: zzl0

Differential Revision: D78272865

fbshipit-source-id: 86bf457accfef35033a8e69732677ab535ac94fc