walkdetector: use "lsof" call to identify walker pid on mac
Summary:
When a "big walk" is first detected, try an lsof-like call on the file path that pushed us over the threshold to (hopefully) discover the PID of the walker. This is not guaranteed to be correct for a myriad of reasons, but in aggregate should be useful.

In practice EdenFS uses NFS on mac, which means EdenFS (and the walk detector) don't know which process is accessing the filesystem. It will be very useful in our metrics if we have the process info.

I tweaked some of the pid related logic to make a non-zero pid "stickier" (i.e. always prefer a non-zero pid when we currently have no pid, and never transition from a non-zero pid back to zero). This helps ensure that we don't accidentally unset the pid, and in general I want to err on the side of having process info, even if we aren't confident in it.

Reviewed By: zzl0

Differential Revision: D78272862

fbshipit-source-id: b83b62d2d1e37012411c3e5b579a47c8afd74eff