Integration test highlighting bug in Mononoke Git shallow fetch --deepen implementation
Summary: As part of investigating S532648, we realized that we end up fetching the same data again and again on incremental fetches of some AOSP repos even though nothing really changes in them. Further investigation revealed a bug in Mononoke's implementation of the `--deepen` variant of shallow fetches. This integration test highlights that problem and will serve as a guard against regression once we fix the problem.

Reviewed By: giorgidze

Differential Revision: D78368863

fbshipit-source-id: fa64ee8c211ce41838fcbbad9777833ab713e79b