backingstore: pause over-prefetching of files
Summary:
Add some smarts so we "pause" a file prefetch if it has prefetched too many files that haven't been accessed. The goal is to avoid very large file prefetches after false positive walk detections, which happens sometimes.

backingstore.walk-prefetch-min-ratio controls the min `accessed/prefetched` ratio, defaulting to 10%. This means we will keep prefetching files as long as 10% of the files we have prefetched are accessed. backingstore.walk-prefetch-max-initial-lag is the `prefetched-accessed` lag we always allow, defaulting to 50k. This needs to be big to allow the prefetch to get going and stay ahead of large walks.

Reviewed By: zzl0

Differential Revision: D78272864

fbshipit-source-id: d74a1725a3afa10456d8786a2f3139d387d77b85