scmstore: always avoid inserting duplicate aux cache entries
Summary:
When inserting into the file (and tree) aux caches after fetching a tree from SLAPI (or CAS), we now always check if the entry exist in the indexedlog before inserting it. This is to avoid duplicate inserts, which will cause the aux cache to roll over prematurely, getting out of sync with the tree cache (which upsets EdenFS performance).

Previously we skipped the read-before-write when fetching from SLAPI assuming that if we are fetching from SLAPI it must be a new tree not in the cache. This is true for the tree part, but still leads to lots of duplicate aux insertions. Imagine a tree with 1000 entries. We fetch the tree, and then cache the 1000 aux entries. Later, a single file in the directory is updated, so we fetch the new tree along with 999 identical entries and one new entry. Previously we would insert all 1000 aux entries again, even though most of them are duplicate. Now we check each one before inserting.

I tested the perf, and it was maybe a tiny bit (2-3%) slower for the initial tree fetch (where we perform a read and then write for each aux entry), but _much_ faster (~70%) for the case of fetching a new "version" of a tree where not many entries changed (because we no longer perform the write).

Reviewed By: MichaelCuevas

Differential Revision: D78381506

fbshipit-source-id: 085f4f501760a29201f19dc25533bccca84c7797