Adding SqlQueryTelemetry to mononoke Transaction
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on tracking Transaction level telemetry and getting repo_ids from the queries.

## This diff

This finishes the work from D78008615, but I split it into a separate diff to make the review easier and provide context.

In order to actually log the data, the Transaction wrapper needs a `SqlQueryTelemetry` object, which has everything needed to instantiate a scuba logger and log.

Because we're currently creating Transaction wrappers by calling `.into()` from the `sql::Transaction`, updating all the callsites will take a while, so I'll do it in multiple diffs.

For now, I'll make it optional, I'll then replace all callsites to instantiate it properly, then will make it required.

Reviewed By: markbt

Differential Revision: D78333807

fbshipit-source-id: 3d87d4a50c1de180995e8a9ece7a37b14449c1d7