Reduce duplication from transaction read queries
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on tracking Transaction level telemetry and getting repo_ids from the queries.

## This diff

The two `query_with_transaction` functions for read queries are **exactly the same**, so let's extract that into a helper.

Reviewed By: markbt

Differential Revision: D78351357

fbshipit-source-id: e3d200fb373a5705e81e74852cdf7c5d201f50c4