Log a vec of repo ids instead of a single one
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on tracking Transaction level telemetry and getting repo_ids from the queries.

## This diff

Change the column we log from repo_id to repo_ids, for a few reasons:

- When we support transaction telemetry in the next few diffs, we'll aggregate telemetry from multiple queries, which means multiple repo ids.
- A single query might involve **multiple repo_ids**, e.g. cross repo tables have source/target or small/large repos.

Reviewed By: markbt

Differential Revision: D78170660

fbshipit-source-id: 0c95d32e5e92b9eca51933ad79de252e4ba1ad0f