Track MysqlQueryTelemetry across a transaction
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on tracking Transaction level telemetry and getting repo_ids from the queries.

## This diff

Start keeping track of some telemetry within a transaction.

- Define a new type to store everything we want to track
- Update the `Transaction` wrapper to store it
- Update the telemetry after every query

To reduce the size of this diff, logging will start happening on the next diff, because I need to start passing `SqlQueryTelemetry` to the `Transaction` wrapper and that will need to be done slowly. I'll explain more in the next diff.

Reviewed By: markbt

Differential Revision: D78008615

fbshipit-source-id: fbf8f442ae6e5880b6503f9d36cebb2993fa6c26