Log query name
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on tracking Transaction level telemetry and getting repo_ids from the queries.

## This diff

Unrelated to the transaction, but it will be useful to get the query that's responsible for a sample.
We'll also be able to quickly group by and drill down on any queries that might cause a SEV.

Reviewed By: markbt

Differential Revision: D78409751

fbshipit-source-id: 8b5b9509981f5dc54eeaf03852891a56c436f287