Use RepositoryId type when defining queries
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on tracking Transaction level telemetry and getting repo_ids from the queries.

## This diff

When queries use `RepositoryId` we log the ids to the SQL metadata table.

Reviewed By: markbt

Differential Revision: D78480982

fbshipit-source-id: fa6255f3ffb0999acc36334c2e4318ae4d2db693