Read/Write the Daemon exit signal
Summary:
Write exit signal for SIGTERM and SIGINT signals for their handlers. In eden startup we should read the `daemon_exit_signal` and log it in the scuba for the silent crash, if the file exist.
Because we only handle SIGTERM and SIGBUS, we only have the signal for these two. SIGKILL and other signals doesn't have this signal, and we record zero in scuba.
As SIGBUS and SIGSEGV are reporting in macos_crash_tailer, we don't need to record them.

Reviewed By: MichaelCuevas

Differential Revision: D78521118

fbshipit-source-id: 18c5b9258eee20d7c48bf42301be5e72bc4d1bc2