Rename tel_logger to sql_query_tel
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on tracking Transaction level telemetry and getting repo_ids from the queries.

## This diff
Lots of variables were named this because of the original name of the struct. Let's rename them to match the new name, SqlQueryTelemetry`.

Reviewed By: lmvasquezg

Differential Revision: D78555009

fbshipit-source-id: b81cb03d8125cc03a5e55e77c3ccb2fe2d7dd1b8