Remove SqlTelemetryLogger arg from transaction queries
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on tracking Transaction level telemetry and getting repo_ids from the queries.

## This diff
Continuing clean up I mentioned in previous diff (D785503640).

All the `query_with_transaction` functions have a `SqlQueryTelemetry` in the transaction object, so no need for them to have that arg.

Reviewed By: lmvasquezg

Differential Revision: D78550994

fbshipit-source-id: 70c675014c52668121b0b20303d6f3f4a20d668d