Stop using `.into()` to get a SqlQueryTelemetry from CoreContext
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on tracking Transaction level telemetry and getting repo_ids from the queries.

## This diff

Follow-up I discussed with markbt (see his comment in D77650966).

`From` is better used for casting between similar/equivalent types (e.g. that are a superset of another). In this case, they are unrelated, so let's use an associated function to build it.

**This diff just finds and replaces callsites that used `.into()` to use the new associated function.**

Files that have relevant changes are:
- fbcode/eden/mononoke/server/context/src/core.rs
- fbcode/eden/mononoke/common/rust/sql_ext/src/mononoke_queries.rs

## Next diffs
In the next few diffs I'll do some more cleanup:

- Rename `tel_logger` to `sql_query_tel` (I updated most, but missed some)
- Remove the `SqlTelemetryLogger` arg from the transaction queries, since the Transaction wrapper always has it.
- Make `SqlTelemetryLogger` required in the non-transaction queries.

Reviewed By: markbt

Differential Revision: D78550364

fbshipit-source-id: 26755546127142c6a07ae1824e3c0b26da58f2c5