Make SqlQueryTelemetry required in the Transaction wrapper
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

This and the next few diffs will be focused on tracking Transaction level telemetry and getting repo_ids from the queries.

## This diff

After updating all the callsites in D78483026, the `Transaction` wrapper should always have a `SqlQueryTelemetry`, so let's make it required and remove all the unnecessary Option checks.

Reviewed By: andreacampi

Differential Revision: D78493134

fbshipit-source-id: e8432b064bf2b7744bee1ec30474b33860204e60