Pass SqlQueryTelemetry to all query calls in sqlblob
Summary:
## This stack
This stack is working on T223577767, i.e. logging telemetry from SQL queries to the metadata XDB.

**This diff and the next few diffs will focus on updating all the `query` callsites to ensure that they pass a `SqlQueryTelemetry`, which is currently an optional arg. This will unblock D78555008, which makes the arg required.**

I'll put up multiple diffs, updating projects or groups of files incrementally, until we can just make the arg required.

## Why do we want this?
If all callsites pass a `SqlQueryTelemetry`, all the queries should have a client correlator logged with it! This will be useful for sev-fighting and leveraging the telemetry we now have in the new `mononoke_sql_telemetry` scuba table.

## This diff

As title

Reviewed By: chiara125

Differential Revision: D78631898

fbshipit-source-id: 1878f851149a13938a739cf5145aff4ee22fc9c8