Fix inefficiencies in object enumeration for shallow clones
Summary:
As part of shallow clones, we are required to enumerate and count the number of distinct blobs and trees. For every commit except the ones in shallow boundary, this is done via GitDeltaManifestv2 which is efficient. For the shallow boundary commits, we use the root `GitTree` and leverage bounded traversal to list all entries in the tree.

This by itself is quite slow but in addition we also fetch the object content itself to get their size which makes the process even slower by introducing the potential for subexecutor problem. This diff fixes that by ensuring we collect  the bounded traversal result and then use bufferring to perform concurrent instead of serial actions.

Reviewed By: YousefSalama

Differential Revision: D78668766

fbshipit-source-id: e7ad8ffd246e27ef60202409f67825ba7e91d9b9