edensparse: call _setupdiff() to override diff operation
Summary:
Currently, when a commit touched files that are not in the eden sparse config, then
`sl diff` will panic with file not found error when diffing between the commit and
working copy .

This diff fixes the issue by using _setupdiff() to overwrite the `remoteworkingfilectx.data()`
and `remoteworkingfilectx.size()` functions. So they fall back to the working copy
parent when the file is missing.

Reviewed By: muirdm

Differential Revision: D78638307

fbshipit-source-id: fd5ba09d012f3eb137d6084e1f8777361a6641b1