delete runSerialPrefetch flag
Summary:
# Project Context
EdenFS has a non-trivial threading model. It has a series of executors, some which have very specific purposes and some which handle generic work. Specific units of work (such as, serving a Thrift call) can jump around from one executor to another. A handful of Thrift entrypoints, like the entrypoints that handle prefetch and Watchman queries, have wide fanout, and can monopolize EdenFS' resources, leading to starvation and eventual Queue Timeouts for other Thrift methods.

Furthermore, there exist foot-guns in EdenFS that have caused repeated SEVs (S412223, S399431, S401889) and have caused user-visible issues like deadlocks and exceptions (D50199539, D41390586).  These issues are easy to accidentally introduce due to the lack of guard rails in place and are difficult to root cause. Much of our threading rules/logic is mandated by in-code comments or learnt prior experience, which is not scalable and makes it hard to onboard and confidently land changes. This project's goal is to de-risk changes that touch these parts of the codebase and to make it easier to understand and optimize EdenFS' resource usage.

# This Diff
This config has been defaulted on, so deleting the consumption code and the fallback path.

This was just added around a config to ensure there wasn't a brutal performance issue if people were relying on the results from prefetch/glob.

We've had no complaints, and a lot of this "serial" code is going to be made first class in the Thrift layer, so we could forward fix if needed to add a "glob threadpool"

# Resources

Reviewed By: MichaelCuevas

Differential Revision: D78850268

fbshipit-source-id: c067227ba16ed3c158cbc8efbb62f0ce215d4fa1