remove dependency on LFS test
Summary:
# This stack
chaosblob is a blobstore abstration for chaos testing which allows for setting fractions of reads and writes that will fail
This feature is only used in a couple of tests, and can easil be replaced. I'm remvoing it for simplicity.

# This diff
In this LFS test we can replace caosblob with readonly, and the result is equivalent: because we fail to write the read will also fail

Reviewed By: chiara125

Differential Revision: D78672848

fbshipit-source-id: f4953ac9dbb8f919baf1be5a721fbda9bd5fc005