remove dependency on walker test
Summary:
# This stack
chaosblob is a blobstore abstration for chaos testing which allows for setting fractions of reads and writes that will fail
This feature is only used in a couple of tests, and can easil be replaced. I'm remvoing it for simplicity.

# This diff
In this walker test we can replace chaosblob with manually deleting a file (blob), and the result is equivalent, even if the error changes slightly

Reviewed By: chiara125

Differential Revision: D78697023

fbshipit-source-id: d0cd56c01da8bc8af6ff7f59a0dc13b6ec5d5d31