Avoid loading objects in shallow read path unless absolutely necessary
Summary:
When enumerating objects as part of shallow fetch or clone, we fetch all the required objects from the blobstore only to get their size. However, this size is then only used for filtering the objects and in most cases there are no filter arguments provided.

This diff takes advantage of that fact and avoids unnecessarily loading objects in memory. This will provide both speed and memory benefits.

Differential Revision: D78805299

fbshipit-source-id: 6052d88cfb30ef3f8ed2fef58c85d28b436464b6