Report errors from building .eden symlinks
Summary: In this diff, we make the symlinks within the <repo_path>/.eden directory a hard requirement for cloning. Previously, if any errors occurred building these symlinks the clone would continue into a bad state. Now the clone command fails early and reports the error back to the Python CLI

Reviewed By: MichaelCuevas

Differential Revision: D78696703

fbshipit-source-id: b021f481c4802503d515d7303046b56d37a00a50