scmstore: optimize writing of file entries to cache
Summary:
Similar to TreeStore, we now buffer the cache entries and flush them to the indexedlog in batches. Since files can be large, I added a 1k limit so we don't buffer indefinitely.

The point of the batch insert is to reduce RwLock contention for the cache operations. Previously we would spawn 4 tasks as we processed the SLAPI results, compressing and writing the indexedlog cache in parallel. This however creates a lot of lock contention even among a single file query (as opposed to other file queries happening at the same time). Batching things up and only acquiring the write lock once smooths things out a lot.

Note that previously the evict_to_cache method would write the SLAPI file data to the indexedlog cache, then immediately query the indexedlog and pass along that mmap'd data (instead of passing along the SLAPI bytes). Presumably the intention was to drop the anon memory as soon as possible, switching to the shared file backed memory. However, two things made this not useful in practice:
1. When writing to indexedlog, normally the entry is first buffered in memory (not backed by mmap). So, reading back the indexedlog entry would usually just be normally memory.
2. Even if we did get mmap'd memory, the file content is compressed in the indexedlog, so we have to copy and allocate normal memory when accessing the file content, anynway.

Now we pass along the SLAPI bytes we already have in hand. This will reduce copies in practice, and allows us to do the batch cache insertions.

Reviewed By: quark-zju

Differential Revision: D78748396

fbshipit-source-id: fa8931da2bdcb903639f456bf83ef25d1ec5c099