walkdetector: store less directory metadata indefinitely
Summary:
Previously we remembered directory metadata forever for big directories (i.e. either a lot of child dirs and/or a lot of files). For very large walks, this can end up retaining a lot of metadata, which makes full scans of the tree slow. Tweak a couple things to reduce cases where we store the metadata forever:
1. Raise "important" *file* threshold to 10x previous value. In general, we are much more concerned about directories with lots of children directories than directories with lots of files due to potential for fast exponential growth as we expand the walk boundary to the next level.
2. Fix an off-by-one error where we were we were saying a directory's large size was "important" even though the size would not have ended us up with an increased walk threshold.

Reviewed By: quark-zju

Differential Revision: D78748394

fbshipit-source-id: aeaf235f64001b725de5258f5bd5985790dd6fc9